ranking forms (#47)
## Notion ticket link
[llsc-74-ranking
forms](https://www.notion.so/uwblueprintexecs/Ranking-Forms-20510f3fb1dc80fd8cfacfd8af1ce2df?source=copy_link)


<!-- Give a quick summary of the implementation details, provide design
justifications if necessary -->
## Implementation description
This PR completes the ranking forms on both the frontend and backend. It
also includes a migration to update the ranking_preferences schema so
that we can store experiences/treatments rather than just qualities, and
so that we can also keep track of whether the scope of the preference
selected is for the volunteer or the volunteer's loved one (which is
needed for cases where the user is a caregiver requesting a caregiver
volunteer). This should hopefully make getting the ranking preferences
easier for the matching part.

A bunch of frontend improvements as well. The main one is the new
protected page component which allows us to restrict certain types of
users from viewing certain pages (so that volunteers can't access the
participant intake or ranking form and participants can't access the
volunteer intake form, etc.).

To complete the forms, we still need to do the secondary application
form for volunteers. I think the backend has been started in another
branch but I don't think we've started the frontend yet.

We also need the logic for controlling what should be shown to users
when they login for both participants and volunteers (eg. should it be
the intake form, intake thank you note, ranking form, ranking thank you
note, or dashboard). This can probably go in a separate PR later? We'll
probably need some new endpoints for allowing admins to move users
between stages and potentially schema changes (haven't thought about
this in much detail yet).

I think some minor styling improvements can be made to better match the
figma designs, as well as making sure our forms look good on mobile as
well.

<!-- What should the reviewer do to verify your changes? Describe
expected results and include screenshots when appropriate -->
## Steps to test
1.


<!-- Draw attention to the substantial parts of your PR or anything
you'd like a second opinion on -->
## What should reviewers focus on?
* 


## Checklist
- [ ] My PR name is descriptive and in imperative tense
- [ ] My commit messages are descriptive and in imperative tense. My
commits are atomic and trivial commits are squashed or fixup'd into
non-trivial commits
- [ ] I have run the appropriate linter(s)
- [ ] I have requested a review from the PL, as well as other devs who
have background knowledge on this PR or who will be building on top of
this PR

---------

Co-authored-by: YashK2005 <yashkoth7@gmail.com>