Reset migrations and created a seeder folder in the backend with the initial experiences/treatments/qualities/roles/forms (#51)
Hack the North pt. 1

Next steps: for the intake forms - updating the seed data with the new
experiences/treatments if needed, fetching the experiences from the
backend rather than hardcoding them in the frontend, having a system so
that cancer patients and caregivers are shown different experiences.

After that I will probably work on the participant dashboard.

The designs for the forms are slightly updated now so that should also
be changed (will leave this for later in the term)

## Notion ticket link
<!-- Please replace with your ticket's URL -->
[Ticket
Name](https://www.notion.so/uwblueprintexecs/Task-Board-db95cd7b93f245f78ee85e3a8a6a316d)


<!-- Give a quick summary of the implementation details, provide design
justifications if necessary -->
## Implementation description
* 


<!-- What should the reviewer do to verify your changes? Describe
expected results and include screenshots when appropriate -->
## Steps to test
1.


<!-- Draw attention to the substantial parts of your PR or anything
you'd like a second opinion on -->
## What should reviewers focus on?
* 


## Checklist
- [ ] My PR name is descriptive and in imperative tense
- [ ] My commit messages are descriptive and in imperative tense. My
commits are atomic and trivial commits are squashed or fixup'd into
non-trivial commits
- [ ] I have run the appropriate linter(s)
- [ ] I have requested a review from the PL, as well as other devs who
have background knowledge on this PR or who will be building on top of
this PR