Add password checks for registration (#54)
## Notion ticket link
No link

## Implementation description
The backend user schema was modified to return an array of requirements
that the password does not meet rather than just raising the first one,
with updated requirements as well. The authApiClient was modified to
handle 422 errors and parse this list of errors. The participant form
frontend was modified check if there is an error before running
`router.push()`, and to loop over the errors to display the requirement
list. In the future we can make this change to admin page as well.

## Steps to test
On the registration page enter a password which does not fit the
requirements. Instead of allowing you to proceed, it should show a
checklist of requirements the password must meet. This should match the
figma design:
https://www.figma.com/design/HhFPmxFewSCBLGkIt9RZen/LLSC-Main-Design-File?node-id=6082-14924&t=xlyQNnm2i6cO0ct6-0

### What it looks like
![Image 2025-09-21 at 11
53 PM](https://github.com/user-attachments/assets/9c6b335b-00d6-44f3-83a5-27e03570124f)

### Backend tests
Tests in `user_schema_test.py` handle password requirement logic and all
of them should pass
![Image 2025-09-22 at 11
18 AM](https://github.com/user-attachments/assets/8c1c24a6-46d0-4512-89a5-88bef6081572)

## What should reviewers focus on?


## Checklist
- [ ] My PR name is descriptive and in imperative tense
- [ ] My commit messages are descriptive and in imperative tense. My
commits are atomic and trivial commits are squashed or fixup'd into
non-trivial commits
- [ ] I have run the appropriate linter(s)
- [ ] I have requested a review from the PL, as well as other devs who
have background knowledge on this PR or who will be building on top of
this PR