LLSC-65: Created form_status column on users table that keeps track o… (#59)
Added form_status column to users table to track onboarding progress.
Added routing in the frontend to automatically take the user to the
correct form/page based on their form_status. This way if i use logs in,
and has already completed the intake form, they won't be shown the
intake form again.

Todo: we need methods to update statuses which will be done as part of
the admin dashboard.
We also need the backend for the volunteer secondary app.

## Notion ticket link
<!-- Please replace with your ticket's URL -->
[Form
Routing](https://www.notion.so/uwblueprintexecs/Intake-Forms-Track-Completion-Status-and-Route-to-Correct-Form-based-on-this-27210f3fb1dc806d8defe3ef1b99eb68?v=27210f3fb1dc8133a6e1000cf4eabc81&source=copy_link)


<!-- Give a quick summary of the implementation details, provide design
justifications if necessary -->
## Implementation description
* 


<!-- What should the reviewer do to verify your changes? Describe
expected results and include screenshots when appropriate -->
## Steps to test
1.


<!-- Draw attention to the substantial parts of your PR or anything
you'd like a second opinion on -->
## What should reviewers focus on?
* 


## Checklist
- [ ] My PR name is descriptive and in imperative tense
- [ ] My commit messages are descriptive and in imperative tense. My
commits are atomic and trivial commits are squashed or fixup'd into
non-trivial commits
- [ ] I have run the appropriate linter(s)
- [ ] I have requested a review from the PL, as well as other devs who
have background knowledge on this PR or who will be building on top of
this PR